move tests to `_qt`
