+++ napari/utils/_tests/test_proxies.py
@@ -1,11 +1,9 @@
-from threading import Thread
 from unittest.mock import patch
 
 import numpy as np
 import pytest
 
 from napari.components.viewer_model import ViewerModel
-from napari.utils import misc
 from napari.utils._proxies import PublicOnlyProxy, ReadOnlyWrapper
 
 
@@ -137,22 +135,3 @@ def test_receive_return_proxy_object():
 def test_viewer_method():
     viewer = PublicOnlyProxy(ViewerModel())
     assert viewer.add_points() is not None
-
-
-def test_proxy_fixture_warning(make_napari_viewer_proxy, monkeypatch):
-    viewer = make_napari_viewer_proxy()
-
-    monkeypatch.setattr(misc, 'ROOT_DIR', '/some/other/package')
-    with pytest.warns(FutureWarning, match='Private attribute access'):
-        viewer.window._qt_window
-
-
-def test_proxy_fixture_thread_error(
-    make_napari_viewer_proxy, single_threaded_executor
-):
-    viewer = make_napari_viewer_proxy()
-    future = single_threaded_executor.submit(
-        viewer.__setattr__, 'status', 'hi'
-    )
-    with pytest.raises(RuntimeError, match='Setting attributes'):
-        future.result()
